who-unfollowed-me/requirements.txt
requests
smtplib
sqlite3
python-dotenv